Use ulong for the %lu / %zu path.
On Linux LP64, unsigned long is 64-bit. Reading it as uint truncates the upper 32 bits for %lu, and this branch also handles %zu when size_t.sizeof == long.sizeof.

+++ stdio/printf.d
@@ -747,7 +747,7 @@ static if (PRINTF_SUPPORT_LONG_LONG) {
 }
           }
           else if (flags & FLAGS_LONG) {
-            uint value = va_arg!uint(va);
+            ulong value = va_arg!ulong(va);
             idx = _ntoa_long(out_fct, buffer, idx, maxlen, value, false, base, precision, width, flags);
           }
           else {
